+++ module/dice.js
@@ -31,6 +31,7 @@ import {RollWithOriginalExpression} from "./roll/roll-with-expression.js";
 export async function d20Roll({parts=[],  partsExpressionReplacements = [], data={}, event={}, rollMode=null, template=null, title=null, speaker=null,
 								  flavor=null, fastForward=null, onClose, dialogOptions, critical=20, fumble=1, targetValue=null,
 								  isAttackRoll=false, options= {}}={}) {
+	critical = critical || 20; //ensure that critical always has a value
 	const rollConfig = {parts, partsExpressionReplacements, data, speaker, rollMode, flavor, critical, fumble, targetValue, isAttackRoll, options }
 
 	// handle input arguments