+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 0.2.54
 - SRD Potions should now automatically target a characters healing surges
 - SRD weapon tables minor size adjustment
+- if no value is entered into a weapons brutal dice value, the system should default to rerolling 1s
+- if no value is entered into a weapons critical hit value, the system should default to use a 20
 
 ## Version 0.2.53
 - Fixed an issue with creating with dragging items into the macro bar that was cause by attempting to use a deprecated call from v8.